*** empty log message ***

+++ Frontal/TPZFrontMatrix.cpp
@@ -252,8 +252,8 @@ void TPZFrontMatrix<store, front>::main()
 template<class store, class front>
 void TPZFrontMatrix<store, front>::CheckCompress()
 {
-	int nfreerate = (int)(fFront.NFree() / fFront.NElements()) * 100;
-	if(nfreerate>20) fFront.Compress();
+	double nfreerate = ( (double)fFront.NFree() / (double)fFront.NElements() ) * 100;
+	if(nfreerate>20.) fFront.Compress();
 }
 
 template<class store, class front>
